@@ -1,10 +1,10 @@
 import React from 'react'
 import Restore from 'react-restore'
-import { isHex } from 'web3-utils'
-import { stripHexPrefix } from 'ethereumjs-util'
+import { isHexString } from '@ethersproject/bytes'
+import { stripHexPrefix } from './../../../../../../resources/utils'
 
 function decodeMessage (rawMessage) {
-  if (isHex(rawMessage)) {
+  if (isHexString(rawMessage)) {
     const buff = Buffer.from(stripHexPrefix(rawMessage), 'hex')
     return buff.length === 32 ? rawMessage : buff.toString('utf8')
   }
