@@ -1,6 +1,5 @@
 import log from 'electron-log'
-import { isValidAddress, addHexPrefix } from 'ethereumjs-util'
-import { Version } from 'eth-sig-util'
+import { isValidAddress, addHexPrefix } from '@ethereumjs/util'
 
 import { AccessRequest, AccountRequest, Accounts, RequestMode, TransactionRequest } from '..'
 import nebulaApi from '../../nebula'
@@ -16,6 +15,7 @@ import provider from '../../provider'
 import { ApprovalType } from '../../../resources/constants'
 
 import reveal from '../../reveal'
+import type { TypedMessage } from '../types'
 
 const nebula = nebulaApi()
 
@@ -510,15 +510,15 @@ class FrameAccount {
     }
   }
 
-  signTypedData (version: Version, typedData: any, cb: Callback<string>) {
-    if (!typedData) return cb(new Error('No data to sign'))
-    if (typeof (typedData) !== 'object') return cb(new Error('Data to sign has the wrong format'))
+  signTypedData (typedMessage: TypedMessage, cb: Callback<string>) {
+    if (!typedMessage.data) return cb(new Error('No data to sign'))
+    if (typeof (typedMessage.data) !== 'object') return cb(new Error('Data to sign has the wrong format'))
     if (this.signer) {
       const s = signers.get(this.signer)
       if (!s) return cb(new Error(`Cannot find signer for this account`))
       const index = s.addresses.map(a => a.toLowerCase()).indexOf(this.address)
       if (index === -1) cb(new Error(`Signer cannot sign for this address`))
-      s.signTypedData(index, version, typedData, cb)
+      s.signTypedData(index, typedMessage, cb)
     } else if (this.smart && this.smart.actor) {
       const actingAccount = this.accounts.get(this.smart.actor)
       if (!actingAccount) return cb(new Error(`Could not find acting account: ${this.smart.actor}`))
@@ -526,7 +526,7 @@ class FrameAccount {
       if (!actingSigner || !actingSigner.verifyAddress) return cb(new Error(`Could not find acting account signer: ${actingAccount.signer}`))
       const index = actingSigner.addresses.map(a => a.toLowerCase()).indexOf(actingAccount.address)
       if (index === -1) cb(new Error(`Acting signer cannot sign for this address, could not find acting address in signer: ${actingAccount.address}`))
-      actingSigner.signTypedData(index, version, typedData, cb)
+      actingSigner.signTypedData(index, typedMessage, cb)
     } else {
       cb(new Error('No signer found for this account'))
     }
