@@ -1,8 +1,9 @@
-import { TypedDataV1, TypedMessage, MessageTypes } from '@metamask/eth-sig-util'
+import { SignTypedDataVersion } from '@metamask/eth-sig-util'
+import type { TypedMessage } from '../accounts/types'
 
-export function getVersionFromTypedData (typedData: TypedDataV1 | TypedMessage<MessageTypes>) {
+export function getVersionFromTypedData (typedData: TypedMessage['data']) {
   if (Array.isArray(typedData)) {
-    return 'V1'
+    return SignTypedDataVersion.V1
   }
 
   const hasUndefinedType = () => typedData.types[typedData.primaryType].some(({ name }) => typedData.message[name] === undefined)
@@ -11,13 +12,13 @@ export function getVersionFromTypedData (typedData: TypedDataV1 | TypedMessage<M
   try {
     // arrays only supported by v4
     if (containsArrays()) {
-      return 'V4'
+      return SignTypedDataVersion.V4
     }
 
     // no v4-specific features so could use either v3 or v4 - default to v4 unless data contains undefined types (invalid in v4)
-    return hasUndefinedType() ? 'V3' : 'V4'
+    return hasUndefinedType() ? SignTypedDataVersion.V3 : SignTypedDataVersion.V4
   } catch (e) {
     // parsing error - default to v4
-    return 'V4'
+    return SignTypedDataVersion.V4
   }
 }
