@@ -1,12 +1,12 @@
 import log from 'electron-log'
 import EventEmitter from 'stream'
-import { addHexPrefix } from 'ethereumjs-util'
+import { addHexPrefix } from '@ethereumjs/util'
 
-import { TransactionData } from '../../../resources/domain/transaction'
 import { deriveHDAccounts } from './derive'
 import crypt from '../../crypt'
-import { TypedData } from 'eth-sig-util'
+import { TransactionData } from '../../../resources/domain/transaction'
 import { getSignerDisplayType } from '../../../resources/domain/signer'
+import type { TypedMessage } from '../../accounts/types'
 
 export interface SignerSummary {
   id: string,
@@ -72,30 +72,30 @@ export default class Signer extends EventEmitter {
   }
 
   open (device?: any) {
-    console.warn('Signer:' + this.type + ' did not implement a open method')
+    log.warn(`Signer: ${this.type} did not implement an open method`)
   }
 
   close () {
-    console.warn('Signer:' + this.type + ' did not implement a close method')
+    log.warn(`Signer: ${this.type} did not implement a close method`)
   }
 
   delete () {
-    console.warn('Signer:' + this.type + ' did not implement a delete method')
+    log.warn(`Signer: ${this.type} did not implement a delete method`)
   }
 
   update (options = {}) {
-    console.warn('Signer:' + this.type + ' did not implement a update method')
+    log.warn(`Signer: ${this.type} did not implement an update method`)
   }
 
   signMessage (index: number, message: string, cb: Callback<string>) {
-    console.warn('Signer:' + this.type + ' did not implement a signMessage method')
+    log.warn(`Signer: ${this.type} did not implement a signMessage method`)
   }
 
   signTransaction (index: number, rawTx: TransactionData, cb: Callback<string>) {
-    console.warn('Signer:' + this.type + ' did not implement a signTransaction method')
+    log.warn(`Signer: ${this.type} did not implement a signTransaction method`)
   }
 
-  signTypedData (index: number, version: string, typedData: TypedData, cb: Callback<string>) {
+  signTypedData (index: number, typedMessage: TypedMessage, cb: Callback<string>) {
     return cb(new Error(`Signer: ${this.type} does not support eth_signTypedData`), undefined)
   }
 }
