@@ -1,14 +1,17 @@
-import { Client, Utils, Constants } from 'gridplus-sdk'
-import { encode } from 'rlp'
-import { padToEven, addHexPrefix } from 'ethereumjs-util'
-import { hexToNumber } from 'web3-utils'
 import log from 'electron-log'
+import { encode } from 'rlp'
+import { Client, Utils, Constants } from 'gridplus-sdk'
+import { padToEven, addHexPrefix } from '@ethereumjs/util'
+import { TypedTransaction } from '@ethereumjs/tx'
+import { SignTypedDataVersion } from '@metamask/eth-sig-util'
+
 import Signer from '../../Signer'
 import { sign, signerCompatibility, londonToLegacy } from '../../../transaction'
-import { TransactionData } from '../../../../resources/domain/transaction'
-import { TypedTransaction } from '@ethereumjs/tx'
 import { Derivation, getDerivationPath } from '../../Signer/derive'
-import { TypedData } from 'eth-sig-util'
+import { hexToInt } from '../../../../resources/utils'
+
+import type { TypedData, TypedMessage } from '../../../accounts/types'
+import type { TransactionData } from '../../../../resources/domain/transaction'
 
 const ADDRESS_LIMIT = 10
 const HARDENED_OFFSET = 0x80000000
@@ -257,15 +260,9 @@ export default class Lattice extends Signer {
     }
   }
 
-  async signTypedData (index: number, version: string, typedData: TypedData, cb: Callback<string>) {
-    const versionNum = (version.match(/[Vv](\d+)/) || [])[1]
-
-    if ((parseInt(versionNum) || 0) < 4) {
-      return cb(new Error(`Invalid version (${version}), Lattice only supports eth_signTypedData version 4+`))
-    }
-
+  async signTypedData (index: number, typedMessage: TypedMessage<SignTypedDataVersion.V4>, cb: Callback<string>) {
     try {
-      const signature = await this.sign(index, 'eip712', typedData)
+      const signature = await this.sign(index, 'eip712', typedMessage.data)
 
       return cb(null, signature)
     } catch (err) {
@@ -339,15 +336,15 @@ export default class Lattice extends Signer {
 
   private createTransaction (index: number, txType: string, chainId: string, tx: TypedTransaction) {
     const { value, to, data, ...txJson } = tx.toJSON()
-    const type = hexToNumber(txType)
+    const type = hexToInt(txType)
 
     const unsignedTx: any = {
       to,
       value,
       data,
       chainId,
-      nonce: hexToNumber(txJson.nonce || ''),
-      gasLimit: hexToNumber(txJson.gasLimit || ''),
+      nonce: hexToInt(txJson.nonce || ''),
+      gasLimit: hexToInt(txJson.gasLimit || ''),
       useEIP155: true,
       signerPath: this.getPath(index)
     }
@@ -361,7 +358,7 @@ export default class Lattice extends Signer {
     optionalFields.forEach(field => {
       if (field in txJson) {
         // @ts-ignore
-        unsignedTx[field] = hexToNumber(txJson[field])
+        unsignedTx[field] = hexToInt(txJson[field])
       }
     })
 
