@@ -1,17 +1,19 @@
-import { rlp, addHexPrefix, stripHexPrefix, padToEven } from 'ethereumjs-util'
-import { TypedData, TypedDataUtils, TypedMessage } from 'eth-sig-util'
 import log from 'electron-log'
-
+import { encode } from 'rlp'
+import { addHexPrefix, stripHexPrefix, padToEven } from '@ethereumjs/util'
+import { SignTypedDataVersion, TypedDataUtils } from '@metamask/eth-sig-util'
 import Transport from '@ledgerhq/hw-transport'
 import Eth from '@ledgerhq/hw-app-eth'
 
 import { Derivation, getDerivationPath, deriveHDAccounts } from '../../Signer/derive'
-import { TransactionData } from '../../../../resources/domain/transaction'
 import { sign } from '../../../transaction'
 import { DeviceError } from '.'
 
+import type { TypedData } from '../../../accounts/types'
+import type { TransactionData } from '../../../../resources/domain/transaction'
+
 export default class LedgerEthereumApp {
-  private eth: Eth;
+  private eth: Eth
 
   constructor (transport: Transport) {
     this.eth = new Eth(transport)
@@ -55,8 +57,8 @@ export default class LedgerEthereumApp {
 
     try {
       const { domain, types, primaryType, message } = TypedDataUtils.sanitizeData(typedData)
-      domainSeparatorHex = TypedDataUtils.hashStruct('EIP712Domain', domain, types).toString('hex')
-      hashStructMessageHex = TypedDataUtils.hashStruct(primaryType as string, message, types).toString('hex')
+      domainSeparatorHex = TypedDataUtils.hashStruct('EIP712Domain', domain, types, SignTypedDataVersion.V4).toString('hex')
+      hashStructMessageHex = TypedDataUtils.hashStruct(primaryType as string, message, types, SignTypedDataVersion.V4).toString('hex')
     } catch (e) {
       throw new DeviceError('Invalid typed data', 99901)
     }
@@ -72,7 +74,7 @@ export default class LedgerEthereumApp {
       // legacy transactions aren't RLP encoded before they're returned
       const message = tx.getMessageToSign(false)
       const legacyMessage = message[0] !== tx.type
-      const rawTxHex = legacyMessage ? rlp.encode(message).toString('hex') : message.toString('hex')
+      const rawTxHex = legacyMessage ? Buffer.from(encode(message)).toString('hex') : message.toString('hex')
 
       return this.eth.signTransaction(path, rawTxHex, null)
     })
