@@ -4,7 +4,7 @@ describe('polygon', () => {
   it('sets the chain id', () => {
     const config = chainConfig(137)
 
-    expect(config.chainIdBN().toNumber()).toBe(137)
+    expect(config.chainId()).toBe(BigInt(137))
   })
 
   it('sets EIP-1559 to be disabled by default', () => {
