@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js'
 import log from 'electron-log'
-import { addHexPrefix } from 'ethereumjs-util'
+import { addHexPrefix } from '@ethereumjs/util'
 
 import {
   addNetwork as addNetworkAction,
