@@ -1,4 +1,4 @@
-const { addHexPrefix } = require('ethereumjs-util')
+const { addHexPrefix } = require('@ethereumjs/util')
 
 export const weiToHex = wei => addHexPrefix(wei.toString(16))
 export const gweiToHex = gwei => weiToHex(gwei * 1e9)
